Show the citation UI in an anchored popover instead of a modal dialog
The "Cite" menu item now opens a popover card anchored to the button that
hosts the menu (the triple-dot button, or the share button in the friendly
UI), following the share menu's PopperCard + click-away pattern. The host
button owns the open state and provides an opener through
OpenCitePopoverContext, since the menu item unmounts when the menu closes.
CitePostDialog is removed.
