+++ packages/lesswrong/components/dropdowns/posts/PostActionsButton.tsx
@@ -1,4 +1,4 @@
-import React, { CSSProperties, useRef, useState } from 'react'
+import React, { CSSProperties, useCallback, useRef, useState } from 'react'
 import MoreHorizIcon from '@/lib/vendor/@material-ui/icons/src/MoreHoriz';
 import MoreVertIcon from '@/lib/vendor/@material-ui/icons/src/MoreVert';
 import { useTracking } from '../../../lib/analyticsEvents';
@@ -10,6 +10,8 @@ import PopperCard from "../../common/PopperCard";
 import PostActions from "./PostActions";
 import LWClickAwayListener from "../../common/LWClickAwayListener";
 import { FeedPostMetaInfo } from '../../ultraFeed/ultraFeedTypes';
+import CitePostPopover from '../../posts/CitePostPopover';
+import { OpenCitePopoverContext } from '../../posts/CitePostPopoverContext';
 
 const styles = defineStyles("PostActionsButton", (theme: ThemeType) => ({
   root: {
@@ -54,6 +56,8 @@ const PostActionsButton = ({post, vertical, popperGap, autoPlace, flip, includeB
   const classes = useStyles(styles);
   const anchorEl = useRef<HTMLDivElement | null>(null);
   const [isOpen, setIsOpen] = useState(false);
+  const [isCiteOpen, setIsCiteOpen] = useState(false);
+  const openCitePopover = useCallback(() => setIsCiteOpen(true), []);
   const {captureEvent} = useTracking();
 
   // This is fine with SSR because the popper will only be rendered after use
@@ -96,9 +100,17 @@ const PostActionsButton = ({post, vertical, popperGap, autoPlace, flip, includeB
     >
       {/*FIXME: ClickAwayListener doesn't handle portals correctly, which winds up making submenus inoperable. But we do still need clickaway to close.*/}
       <LWClickAwayListener onClickAway={() => handleSetOpen(false)}>
-        <MenuComponent post={post} closeMenu={() => handleSetOpen(false)} includeBookmark={includeBookmark} onSeeLess={onSeeLess} isSeeLessMode={isSeeLessMode} postMetaInfo={postMetaInfo} />
+        <OpenCitePopoverContext.Provider value={openCitePopover}>
+          <MenuComponent post={post} closeMenu={() => handleSetOpen(false)} includeBookmark={includeBookmark} onSeeLess={onSeeLess} isSeeLessMode={isSeeLessMode} postMetaInfo={postMetaInfo} />
+        </OpenCitePopoverContext.Provider>
       </LWClickAwayListener>
     </PopperCard>
+    <CitePostPopover
+      post={post}
+      anchorEl={anchorEl.current}
+      open={isCiteOpen}
+      onClose={() => setIsCiteOpen(false)}
+    />
   </div>
 }
