+++ packages/lesswrong/components/posts/SharePostButton.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from 'react';
+import React, { useCallback, useRef, useState } from 'react';
 import { registerComponent } from '../../lib/vulcan-lib/components';
 import withErrorBoundary from '../common/withErrorBoundary';
 import classNames from 'classnames';
@@ -10,6 +10,8 @@ import ForumIcon from "../common/ForumIcon";
 import PopperCard from "../common/PopperCard";
 import LWClickAwayListener from "../common/LWClickAwayListener";
 import SharePostActions from "../dropdowns/posts/SharePostActions";
+import CitePostPopover from "./CitePostPopover";
+import { OpenCitePopoverContext } from "./CitePostPopoverContext";
 import { defineStyles } from '@/components/hooks/defineStyles';
 import { useStyles } from '@/components/hooks/useStyles';
 
@@ -33,6 +35,8 @@ const SharePostButton = ({post, className}: {
   const classes = useStyles(styles);
   const anchorEl = useRef<HTMLDivElement | null>(null)
   const [isOpen, setIsOpen] = useState<boolean>(false)
+  const [isCiteOpen, setIsCiteOpen] = useState(false);
+  const openCitePopover = useCallback(() => setIsCiteOpen(true), []);
   const { captureEvent } = useTracking()
   
   const shareClicked = () => {
@@ -69,9 +73,18 @@ const SharePostButton = ({post, className}: {
       allowOverflow
     >
       <LWClickAwayListener onClickAway={() => setIsOpen(false)}>
-        <SharePostActions post={post} onClick={() => setIsOpen(false)} />
+        <OpenCitePopoverContext.Provider value={openCitePopover}>
+          <SharePostActions post={post} onClick={() => setIsOpen(false)} />
+        </OpenCitePopoverContext.Provider>
       </LWClickAwayListener>
     </PopperCard>
+    <CitePostPopover
+      post={post}
+      anchorEl={anchorEl.current}
+      open={isCiteOpen}
+      onClose={() => setIsCiteOpen(false)}
+      placement="bottom"
+    />
   </div>
 }
